requirements.txt
selenium>=4.15.0
python-dotenv>=1.0.0
pypdf>=4.0.0
pyautogui>=0.9.54
opencv-python-headless>=4.8.0
pyperclip>=1.8.2
